moe: share the MoE building blocks between loaders; crate-wide warning cleanup (#66)
The qwen3moe and deepseek2 loaders had grown the same code side by side:
the per-layer KV cache and its truncation, the causal mask, the routed
expert dispatch (prefill bucketing, the serial decode path, the
host/device activation hop from #60) and the slicing of a stacked
`[n_expert, out, in]` tensor into per-expert weights (borrow from the
mapping, upload from the mapping, streamed fallback).  New `src/moe.rs`
holds the one copy:

* `is_routed_expert` (was duplicated in engine.rs and deepseek4.rs),
  `KvCache`, `truncate_kv(.., seq_dim)`, `causal_mask`, `on_device`.
* `Expert` trait + `dispatch_prefill` / `dispatch_decode`.  The router
  out-of-range check qwen3moe had now covers deepseek2 too (it would have
  panicked on the slice index), and drops the stray `eprintln!`.
* `ExpertTensor::{lookup, borrow_host, upload_from_mmap, read_and_split}`.
  qwen3moe keeps only its zero-copy Metal and paged variants around them.
* `ExpertHandles::from_parts` replaces the three-way match in both
  `load_moe`s; `Mlp::prefetch` is `ExpertHandles::prefetch`.

Both loaders' `forward` borrow `Shared` instead of cloning the `Arc`
(the clone was never needed; the fields are disjoint).  deepseek4 uses
the shared hop and predicate; the engine's two `probed_*_names` fold into
one helper.  Net -380 lines with the new module's tests included.

Warning cleanup across the crate, examples and tests (clippy is clean
apart from one `unknown lint` on a version-dependent allow): dead
`Mlp::prefetch` and `FeedForward::forward` in deepseek2/qwen3moe, unused
imports and `mut`s, `is_multiple_of`, `first()`, a duplicated test line,
an unused type parameter, unchecked `Result`s in bench_qwen3moe, a
non-snake-case local, `const` thread-locals.  The batched decode loop
from #54 (`forward_tokens_batched`, `SeqDecodeState`,
`decode_loop_batched`) has no caller yet; it is annotated as pending
its `JOSHUA_BATCH` mode rather than removed.

Tests: new unit tests in `moe.rs` (prefill == decode, out-of-range ids,
truncation on either layout, mask offsets, the three expert-slicing
paths agree with candle's reading).  `cargo test --no-fail-fast` is
unchanged from main: everything passes except the three pre-existing
deepseek4 mmap-vs-streamed failures.

+++ examples/bench_mmap.rs
@@ -4,7 +4,7 @@
 //! Usage: cargo run --release --example bench_mmap -- <model.gguf>
 
 use std::fs::File;
-use std::io::{Cursor, Read, Seek};
+use std::io::{Cursor, Read};
 use std::sync::Arc;
 use std::time::Instant;
 
@@ -104,7 +104,7 @@ fn main() {
         aqm.forward(&axs).unwrap();
     });
 
-    stream_probe::run(Arc::clone(&mmap), base, per_bytes, ne as usize);
+    stream_probe::run(Arc::clone(&mmap), base, per_bytes, ne);
 }
 
 // ── appended: raw streaming probe over expert regions ────────────────────
